Package classes into 'sana'
