+++ src/main/java/sana/AddCommand.java
@@ -1,3 +1,4 @@
+package sana;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
