+++ src/main/java/sana/Command.java
@@ -1,3 +1,4 @@
+package sana;
 public abstract class Command {
 
     protected String cmd;
