+++ src/main/java/sana/DeleteCommand.java
@@ -1,3 +1,4 @@
+package sana;
 public class DeleteCommand extends Command {
 
     public DeleteCommand(String cmd, String arguments) {
@@ -5,8 +6,12 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        Task deletedTask = tasks.delete(Integer.parseInt(arguments) - 1);
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws SanaException {
+        if (arguments.isBlank()) {
+            throw new SanaException("Error! Need to specify which task to delete");
+        }
+
+        Task deletedTask = tasks.delete(Integer.parseInt(arguments));
         System.out.println("Noted. I've removed this task:\n" + deletedTask.toString() + "\n"
                 + "Now you have " + tasks.size() + (tasks.size() <= 1 ? " task" : " tasks")
                 + " in the list");
