+++ src/main/java/sana/Event.java
@@ -1,3 +1,4 @@
+package sana;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
