+++ src/main/java/sana/Sana.java
@@ -1,3 +1,4 @@
+package sana;
 public class Sana {
 
     private Storage storage;
