+++ src/main/java/sana/SanaException.java
@@ -1,3 +1,4 @@
+package sana;
 public class SanaException extends Exception {
 
     public SanaException(String errorMessage) {
