+++ src/main/java/sana/Storage.java
@@ -1,3 +1,4 @@
+package sana;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
