+++ src/main/java/sana/Ui.java
@@ -1,3 +1,4 @@
+package sana;
 import java.util.Scanner;
 
 public class Ui {
